Attempt at fixing test case failure: testTwoTasks(at.silverstrike.pcc.test.gtaskprioritycalculator.TestDefaultGoogleTasksPriorityCalculator)  Time elapsed: 0.523 sec  <<< ERROR! java.lang.NullPointerException 	at at.silverstrike.pcc.impl.gtaskprioritycalculator.DefaultGoogleTasksPriorityCalculator.setPriorities(DefaultGoogleTasksPriorityCalculator.java:46) 	at at.silverstrike.pcc.impl.gtaskprioritycalculator.DefaultGoogleTasksPriorityCalculator.run(DefaultGoogleTasksPriorityCalculator.java:38) 	at at.silverstrike.pcc.test.gtaskprioritycalculator.TestDefaultGoogleTasksPriorityCalculator.testTwoTasks(TestDefaultGoogleTasksPriorityCalculator.java:65)

+++ src/main/java/at/silverstrike/pcc/api/gtaskprioritycalculator/GoogleTasksPriorityCalculator.java
@@ -24,5 +24,5 @@ public interface GoogleTasksPriorityCalculator extends SingleActivityModule {
             setTasks(
                     final Map<String, com.google.api.services.tasks.v1.model.Task> aTasksByTaskIds);
 
-    Map<String, Long> getPrioritiesByTaskIds();
+    Map<String, Integer> getPrioritiesByTaskIds();
 }
